Rename obs-ndi to DistroAV
It appears that OBS requested that obs-ndi change their name.  Their
new name is DistroAV.  We'll rename the package here as well.  This
should be seamless, portage should try to update old references for
users.  This may require an etc-update to change the name for
accept_keywords and sets files as needed.

Thanks to kripton for the heads-up.

Signed-off-by: bunder2015 <omfgbunder@gmail.com>